Fix "Conn close because of keep_alive_timeout" when reusing  `Seahorse::Client::NetHttp::ConnectionPool` sessions
Currently, the `http_idle_timeout` is described as "The number of seconds a connection is allowed to sit idble before it is considered stale" with a default value of 5 seconds [1]. These sessions are keeping a reference to the opened `Net::HTTP` connections that have `keep_alive_timeout` defaulting to 2 seconds [2][3].

The issue here is that any session in the pool reused after the 2 seconds mark will result in an automatic reconnect (i.e. re-establishing a TCP session). This means that setting `http_idle_timeout` to any value greater than 2 seconds is irrelevant.

Setting the `http.keep_alive_timeout` to equals the configured `http_idle_timeout` should allow for all sessions in the pool to be readily used upon checkout.

[1] https://docs.aws.amazon.com/sdk-for-ruby/v3/api/Aws/S3/Client.html
[2] https://github.com/ruby/ruby/blob/af11efd377965b6601bb54aa79072ef0789dc525/lib/net/http.rb#L668
[3] https://ruby-doc.org/stdlib-2.7.0/libdoc/net/http/rdoc/Net/HTTP.html#keep_alive_timeout-attribute-method