Fix ineffective option `http_idle_timeout` in `Seahorse::Client::NetHttp::ConnectionPool` (#2206)
* Fix numerous typos in `Seahorse::Client::NetHttp::ConnectionPool`

* Fix endpoint param doc and example for `Seahorse::Client::NetHttp::ConnectionPool#request`

* Fix erroneously reaped sessions from `Seahorse::Client::NetHttp::ConnectionPool` due to bad `last_used` time calculation

To detect which sessions are stale and remove them from the pool the `_clean` method is comparing the current and `session.last_used` timestamps. If the difference is larger than the predefined `http_idle_timeout` the session is considered stale and is terminated.

The issue here is that the `session.last_used` value is set to be the time right before a request is made thus wrongfully including the request duration in the idle time equation consequently terminating the session instead of reusing it. This is often manifested when a slow connection causes a request (e.g. uploaded part) to take longer than the configured `http_idle_timeout`.

Moving the time sampling after the request is complete fixes the issue and is in line with how Ruby `Net::HTTP` is doing its internal `keep_alive_timeout` calculations [1].

[1] https://github.com/ruby/ruby/blob/af11efd377965b6601bb54aa79072ef0789dc525/lib/net/http.rb#L1598

* Use monotonic clocks when reaping sessions in `Seahorse::Client::NetHttp::ConnectionPool`

Improves the accuracy of `session.last_used`/`http_idle_timeout` diff calculation by using `Process.clock_gettime(Process::CLOCK_MONOTONIC, :millisecond)
`instead of `Time.now` [1][2]. This is in line with how Ruby `Net::HTTP` is doing its internal `keep_alive_timeout` calculations [3].

[1] https://www.rubydoc.info/stdlib/core/Process:clock_gettime
[2] https://blog.dnsimple.com/2018/03/elapsed-time-with-ruby-the-right-way/
[3] https://github.com/ruby/ruby/blob/af11efd377965b6601bb54aa79072ef0789dc525/lib/net/http.rb#L1598

* Fix "Conn close because of keep_alive_timeout" when reusing  `Seahorse::Client::NetHttp::ConnectionPool` sessions

Currently, the `http_idle_timeout` is described as "The number of seconds a connection is allowed to sit idble before it is considered stale" with a default value of 5 seconds [1]. These sessions are keeping a reference to the opened `Net::HTTP` connections that have `keep_alive_timeout` defaulting to 2 seconds [2][3].

The issue here is that any session in the pool reused after the 2 seconds mark will result in an automatic reconnect (i.e. re-establishing a TCP session). This means that setting `http_idle_timeout` to any value greater than 2 seconds is irrelevant.

Setting the `http.keep_alive_timeout` to equals the configured `http_idle_timeout` should allow for all sessions in the pool to be readily used upon checkout.

[1] https://docs.aws.amazon.com/sdk-for-ruby/v3/api/Aws/S3/Client.html
[2] https://github.com/ruby/ruby/blob/af11efd377965b6601bb54aa79072ef0789dc525/lib/net/http.rb#L668
[3] https://ruby-doc.org/stdlib-2.7.0/libdoc/net/http/rdoc/Net/HTTP.html#keep_alive_timeout-attribute-method