allow default ws and wss host:ports to work
Before this commit, it was not possible to presign a websocket url. The
method to determine if the url was standard only took into account
http:80 and https:443 as standard. If a wss url was provided for
presign, it resulted in the signed header host including a trailing
colon and prevented the signautre from working.

For example, before this commit, a call to presign_url with url:
'wss://domain.com', would result in a signed header host of:
'domain.com:', which would prevent the signature from working.