handle known and unknown URI scheme requests
Before this commit, it was not possible to presign a URL other than
HTTP/HTTPS, unless an explict port was provided. The discovered use case
in the commit was for WebSocket (WS/WSS) presigned URLs.

If a URL other than HTTP/HTTPS was provided for presign, and it didn't
include an explicit port, it resulted in the signed header host
including a trailing colon and prevented the signautre from working.

For example, before this commit, a call to presign_url with url:
'wss://domain.com', would result in a signed header host of:
'domain.com:', which prevented the signature from working.