explicit test of known uri with no provided port
Explicitly state and test that Host is correct when a known URI scheme
is used and port is not explicitly provided. The changed #host logic
depends on the fact that uri.default_port and uri.port are set by URI
library and will be equal. This test ensures that to be the case; i.e
that we don't inadvertantly return "#{uri.host}:#{uri.port}".