Presign a WebSocket URL - wss://domain.com/mqtt (#2232)
* allow default ws and wss host:ports to work

Before this commit, it was not possible to presign a websocket url. The
method to determine if the url was standard only took into account
http:80 and https:443 as standard. If a wss url was provided for
presign, it resulted in the signed header host including a trailing
colon and prevented the signautre from working.

For example, before this commit, a call to presign_url with url:
'wss://domain.com', would result in a signed header host of:
'domain.com:', which would prevent the signature from working.

* handle known and unknown URI scheme requests

Before this commit, it was not possible to presign a URL other than
HTTP/HTTPS, unless an explict port was provided. The discovered use case
in the commit was for WebSocket (WS/WSS) presigned URLs.

If a URL other than HTTP/HTTPS was provided for presign, and it didn't
include an explicit port, it resulted in the signed header host
including a trailing colon and prevented the signautre from working.

For example, before this commit, a call to presign_url with url:
'wss://domain.com', would result in a signed header host of:
'domain.com:', which prevented the signature from working.

* explicit test of known uri with no provided port

Explicitly state and test that Host is correct when a known URI scheme
is used and port is not explicitly provided. The changed #host logic
depends on the fact that uri.default_port and uri.port are set by URI
library and will be equal. This test ensures that to be the case; i.e
that we don't inadvertantly return "#{uri.host}:#{uri.port}".