aws-sdk-core/xml/builder: remove default xml indentation (#2527)
XML document indentation increases the body size. It can
be useful for debugging purposes, but under normal operations,
it is not helpful.

The result of a larger body size makes operations like
`complete_multipart_upload` fail for certain S3 gateways
(Ceph Object Gateway at least) when having 10k parts,
because of the body being too large.

Compared to boto3, boto3 does not indent, which makes
the body smaller than the fatal limit.

This commit removes the default indentation to have the
same behavior as boto3 and to be able to use multipart
uploads with 10k parts with Ceph Object Gateway.