Refactor PageableResponse to avoid busting Ruby's constant cache
Ruby has a global constant cache, which is basically a counter
that is incremented anytime a constant is set or modified.

Whenever it is bumped, the vast majority of the Ruby VM inline caches
are invalidated, as well as JITed code. So busting this cache
at "runtime" is very detrimental to performance.

`Object#extend` doesn't normally bust the cache, except if the
module contains constants of it's own:

```ruby
module A; end
obj.extend(A) # The global cache is preserved

module B
  FOO = 1
end
obj.extend(B) # The global cache is invalidated
```

In this case, the constant causing the invalidation is
`Aws::PageableResponse::LastPageError`.

So to avoid regularly flushing the cache, this PR move all
pagination methods to a submodule, and include this one.