CredentialProviderChain: Use Aws.shared_config when deciding on region
The current implementation ignores any data that is retrieve through
user configuration schema. By calling Aws.shared_config.region, it
allows the call to attempt to retrieve the region from any configuration
elsewhere in the system that wasn't provided in the configuration
object.

As this is a fallback, this should have no implication on working system
that already rely on the config object.