Update checksums.rb
Hi!

Thank you for your great job!

Headers are missing :complete_multipart_upload so I need to calculate it outside the plugin. I don't know if this is intended but it is the oposite of what I understand by ready this file comments.

By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.