Fix flaky auth_token spec caused by time zone sensitivity
The `RDS::AuthTokenGenerator` spec was written in a way that assumed
the system clock is always UTC. This meant that the spec would always
pass in CI, since the GitHub Actions runner uses UTC, but sometimes fail
when running locally.

Specifically, after 4pm PST, the local date and UTC dates are different,
leading to a mismatch in the expected and actual date.

This commit explicitly converts the date to UTC with `.utc` so that the
spec now consistently passes regardless of local time zone.