fix_multipart_copy_sse
When copying objects using `Aws::S3::Object.copy_to(..., multipart_copy: true)`,
the `ObjectMultipartCopier` was inferring the metadata of the destination object
from that of the source object. This is fine for `content_type`, `content_length`, etc.;
however, the `ssekms_key_id` and `server_side_encryption` attributes were also copied.
This causes the destination object to use the kms key of the source, which is not usually
appropriate if the destination bucket has a different default encryption than the source.
This is especially true when copying from one account to the other.