Improve code generation to avoid unused variable warnings
The Rails test suite had these warnings for ages, which I'd like
to fix:

```
aws-sdk-s3/endpoint_provider.rb:22: warning: assigned but unused variable - key
aws-sdk-s3/endpoint_provider.rb:24: warning: assigned but unused variable - copy_source
aws-sdk-s3/endpoint_provider.rb:23: warning: assigned but unused variable - prefix
```

Ultimately I don't think pre-assigning to variables is really worth it,
as the `parameters` argument is a T_STRUCT so access is really fast.

But I suspect this was done to make the code generation simpler.
In the end this adds a little bit of complexity, but not that much.